requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
psycopg2
python-dotenv

# ----------------------
# Production Server
# ----------------------
gunicorn
